* Show version count in the TableView of the restore explorer (meaning how many versions of a file have been backed up).
* If a restore is attempted on a file that was recently backed up, the restore may fail because the file hasn't shown up in the gmail account
due to mail delays.  Instead of printing a generic error, we say something like "the file was recently backed up and may not be available
for several hours, please try again later."

* During a backup operation, during the initial phase of scanning all the files in the BackupSpecification (this is the estimate operation),
previously we were only updating the progress dialog with the files that were in need of backup.  This means that if you were asking to
backup a huge directory where only a few files were dirty, you wouldn't get much feedback.  Now we update the progress dialog
every time a file is scanned whether it is dirty or not (only for the estimate).  During backup, of course, we only show the files that
are being backed up.

* Use MD5 digests of file contents to :
	a) Generate GUIDs
	b) Detect more accurately whether files need backup.  Before we'd backup a file if its lastMod time was changed.  Now we check the
		MD5 digest to make sure it has changed.
	c) Allow two files with the same content but different names/locations to share the same email backup message.  So if c:/foo.txt is the
		same as c:/biff/bar/baz/other.cfg, only one email will be sent.  The RevisionIdentifiers do ref counting so we know don't delete the
		email until all refs are gone.

* A new serialization scheme that is smokin' fast and very compact relative to our old DataOutputStream scheme, or netscape IFC, or java
serialization.  Compared to our old scheme, its less than half the size, and is twice as fast.  Compared to IFC, its about 2/3rds the size and about
2.5x faster.  Compared to java serialization its about 1/3rd the size, and about 4x as fast.  It is also much more flexible than our original cheazy
DataOutputStream.  It supports shared references, class versioning, and various optimizations like archiving common strings, writing compact
ints/longs, and optimized references based on whether there are multiple holders of the reference or whether the reference is of fixed type or
not.  See com.toubassi.archive.

* If a backup estimate returns that zero files need a backup, we should immediately say no files needed a backup, rather than doing the backup
which just delays the inevitable).

+++ src/com/toubassi/filebunker/vault/Vault.java
@@ -99,16 +99,7 @@ public class Vault implements NotificationListener, XMLSerializable
             NotificationCenter.sharedCenter().register(store, this);
 
             // Create the index
-            File databaseFile = new File(configDirectory, "database");
-        
-            backupdb = new BackupDatabase();
-
-            long start, end;
-
-            start = System.currentTimeMillis();
-            backupdb.setFile(databaseFile);
-            end = System.currentTimeMillis();
-            //System.out.println("Load time: " + (end - start));
+            backupdb = new BackupDatabase(configDirectory);
 
             NotificationCenter.sharedCenter().register(backupdb, this);
         }
@@ -200,9 +191,9 @@ public class Vault implements NotificationListener, XMLSerializable
     public BackupEstimate estimateBackup(BackupSpecification spec,
             FileOperationListener listener) throws VaultException
     {
-        BackupEstimate estimate = new BackupEstimate(listener, backupdb);
+        BackupEstimate estimate = new BackupEstimate(null, backupdb);
 
-        spec.find(backupdb, estimate);
+        spec.find(backupdb, estimate, listener);
 
         return estimate;
     }
@@ -212,12 +203,22 @@ public class Vault implements NotificationListener, XMLSerializable
     {
         try {
             file = file.getCanonicalFile();
+            
             RevisionIdentifier identifier;
-            long[] size = new long[1];
+            FileDigest digest = new FileDigest(file);
 
-            identifier = store.backupFile(file, null, size, listener);
+            // See if there is an existing file in the backup repository
+            // that has the same content.
+            FileRevision existingRevision = backupdb.fileRevisionWithDigest(digest);
+            if (existingRevision != null) {
+                identifier = existingRevision.identifier();
+            }
+            else {
+                identifier = new RevisionIdentifier(digest, file.length());
+                store.backupFile(file, null, identifier, listener);
+            }
 
-            return backupdb.recordRevision(file, date, identifier, size[0]);
+            return backupdb.recordRevision(file, date, identifier);                
         }
         catch (IOException e) {
             throw new VaultException(file, e);
@@ -275,7 +276,10 @@ public class Vault implements NotificationListener, XMLSerializable
                         listener.willProcessFile(backupdb.file());
                     }
                     
-                    store.backupFile(backupdb.file(), "BackupIndex", new long[1], listener);
+                    FileDigest digest = new FileDigest(backupdb.file());
+                    long size = backupdb.file().length();
+                    RevisionIdentifier identifier = new RevisionIdentifier(digest, size);
+                    store.backupFile(backupdb.file(), "BackupIndex", identifier, listener);
                 }
             }
             catch (VaultException e) {
@@ -454,11 +458,20 @@ public class Vault implements NotificationListener, XMLSerializable
         return directory;        
     }
 
-    private void deleteRevision(FileRevision revision) throws VaultException,
+    /**
+     * Returns true if any data was removed from the store.  If multiple
+     * revisions are sharing the same data in the store (2 files with the
+     * same content) then this may amount to an "unlink" of a file with
+     * multiple hard links.
+     */
+    private boolean deleteRevision(FileRevision revision) throws VaultException,
             IOException
     {
-        store.deleteFile(revision.identifier(), revision.backedupSize());
-        backupdb.removeRevision(revision);
+        if (backupdb.removeRevision(revision)) {
+            store.deleteFile(revision.identifier(), revision.backedupSize());
+            return true;
+        }
+        return false;
     }
 
     private void performDelete(File file, Date date) throws VaultException,
@@ -484,7 +497,6 @@ public class Vault implements NotificationListener, XMLSerializable
         } else {
             FileRevision fileRevision = (FileRevision) revision;
 
-            System.out.println("Deleting " + file);
             deleteRevision(fileRevision);
         }
     }
